# openapi.yaml
openapi: 3.0.3
info:
  title: AKSV Volunteer Platform API
  version: 1.0.0
  description: API contract for Users (Volunteers) and Events resources on AKSV platform.
servers:
  - url: https://api.aksvplatform.com/v1
paths:
...
  /events/{eventId}/join:
    post:
      summary: Volunteer joins an event and selects time slots
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventJoinRequest"
      responses:
        "200":
          description: Volunteer joined event successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventJoinResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Conflict - time slot already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  schemas:
...
    EventJoinRequest:
      type: object
      properties:
        userId:
          type: string
          example: "user123"
        selectedTimeSlots:
          type: array
          items:
            type: string
            example: "09:00-11:00"
      required:
        - userId
        - selectedTimeSlots
    EventJoinResponse:
      type: object
      properties:
        message:
          type: string
          example: "Volunteer joined event successfully."
        eventId:
          type: string
          example: "event123"
        userId:
          type: string
          example: "user123"
        selectedTimeSlots:
          type: array
          items:
            type: string
            example: "09:00-11:00"
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Time slot already taken."
  responses:
    BadRequest:
      description: Bad request due to invalid input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"